Backend/requirements.txt
fastapi
httpx
python-dotenv
uvicorn 
python-multipart